F! ECS & RNG uses uints as entity indices & primary data types

+++ Assets/Scripts/ComputeShaderEngine.cs
@@ -13,10 +13,10 @@ public class ComputeShaderEngine : MonoBehaviour
 
     [Header("Random numbers")] 
     public bool RngSeedFromTimer = false;
-    public int RngSeed = 0;
-    public int RngMaxUniform = 1000000;
-    public int RngCount = 256;
-    public int RngStateLength = 55;
+    public uint RngSeed = 0;
+    public uint RngMaxUniform = 1000000;
+    public uint RngCount = 256;
+    public uint RngStateLength = 55;
     
     private ComputeBuffer _rngState = null;
     
@@ -29,29 +29,29 @@ public class ComputeShaderEngine : MonoBehaviour
         }
         else
         {
-            Random.InitState( RngSeed );
+            Random.InitState( (int)RngSeed );
         }
         
-        int[] rngStateData = new int[RngCount*(RngStateLength+1)];
+        uint[] rngStateData = new uint[RngCount*(RngStateLength+1)];
         for (int i = 0; i < RngCount; i++)
         {
             rngStateData[i * (RngStateLength + 1)] = 0;
             for (int j = 0; j < RngStateLength; j++)
             {
-                rngStateData[i * (RngStateLength + 1) + j + 1] = Random.Range( 0, RngMaxUniform );
+                rngStateData[i * (RngStateLength + 1) + j + 1] = (uint)Random.Range( 0, (int)RngMaxUniform );
             }
         }
-        _rngState = new ComputeBuffer( RngCount*(RngStateLength+1), sizeof(int) );
+        _rngState = new ComputeBuffer( (int)(RngCount*(RngStateLength+1)), sizeof(int) );
         _rngState.SetData( rngStateData );
         
-        ComputeShader.SetInt( "_rngMax", RngMaxUniform);
-        ComputeShader.SetInt( "_rngCount", RngCount);
-        ComputeShader.SetInt( "_rngStateLength", RngStateLength);
+        ComputeShader.SetInt( "_rngMax", (int)RngMaxUniform);
+        ComputeShader.SetInt( "_rngCount", (int)RngCount);
+        ComputeShader.SetInt( "_rngStateLength", (int)RngStateLength);
 
         ComputeShaderEmulator._rngMax = RngMaxUniform;
         ComputeShaderEmulator._rngCount = RngCount;
         ComputeShaderEmulator._rngStateLength = RngStateLength;
-        ComputeShaderEmulator._rngState = new int[rngStateData.Length];
+        ComputeShaderEmulator._rngState = new uint[rngStateData.Length];
         rngStateData.CopyTo(ComputeShaderEmulator._rngState, 0);
     }
